Sudoku Solver
This is a code written in Python that will act as a Sudoku Solver. This program makes use of a number of topics of DSA such as backtracking (recursion) as well as the usual properties of array. By doing this project, I do hope to learn & implement more & more cryptic DSA problems in Python.
P.S.: Sudoku is a number game made up of a grid of square matrices, for instance, nine 3x3 matrices in a 9x9 matrix (these can be found in newspapers and on various edu-gaming websites too). The rule is simple, no same number should exist in the same row, column and grid. All numbers need to be unique. 
Now since, ours is a program, a number of other conditions also need to be kept in mid. These conditions have been explained in the form of comments, alongside the corresponding code, as and when needed.