Create DFS_Code
Based on the Depth First Search [DFS] Algorithm. 
The algorithm of DFS can be used for both directed as well as non-directed graphs. It can also be used for tree structures.
The algorithm starts at the root (top) node of a tree and goes as far as it can down a given branch (path), then backtracks until it finds an unexplored path, and then explores it.